+++ golang/extra_boost/extra_boost_lib/go.mod
@@ -1,0 +1,9 @@
+module example.com/extra_boost_lib
+
+go 1.16
+
+require (
+	github.com/sbinet/npyio v0.5.2
+	gonum.org/v1/gonum v0.9.1
+	gorgonia.org/tensor v0.9.20
+)
