+++ golang/extra_boost/extra_boost_main/go.mod
@@ -1,0 +1,7 @@
+module extra_boost_main
+
+go 1.16
+
+replace example.com/extra_boost_lib => ../extra_boost_lib
+
+require example.com/extra_boost_lib v0.0.0-00010101000000-000000000000
